Move validation function to cs.rs
