+++ nalgebra-sparse/src/cs.rs
@@ -1,3 +1,4 @@
+use std::iter::FromIterator;
 use std::mem::replace;
 use std::ops::Range;
 
@@ -6,7 +7,7 @@ use num_traits::One;
 use nalgebra::Scalar;
 
 use crate::pattern::SparsityPattern;
-use crate::{SparseEntry, SparseEntryMut};
+use crate::{SparseEntry, SparseEntryMut, SparseFormatError, SparseFormatErrorKind};
 
 /// An abstract compressed matrix.
 ///
@@ -542,4 +543,134 @@ pub fn convert_counts_to_offsets(counts: &mut [usize]) {
         *i_offset = offset;
         offset += count;
     }
+}
+
+/// Validates cs data, sorts minor indices and values
+pub(crate) fn validate_and_optionally_sort_cs_data<T>(
+    major_dim: usize,
+    minor_dim: usize,
+    major_offsets: &[usize],
+    minor_indices: &mut [usize],
+    values: Option<&mut [T]>,
+    sort: bool,
+) -> Result<(), SparseFormatError>
+where
+    T: Scalar,
+{
+    if sort && minor_indices.len() > 0 && values.is_none() {
+        return Err(SparseFormatError::from_kind_and_msg(
+            SparseFormatErrorKind::InvalidStructure,
+            "No values provided for sorting.",
+        ));
+    }
+    if major_offsets.len() == 0 {
+        return Err(SparseFormatError::from_kind_and_msg(
+            SparseFormatErrorKind::InvalidStructure,
+            "Number of offsets should be greater than 0.",
+        ));
+    }
+    if major_offsets.len() != major_dim + 1 {
+        return Err(SparseFormatError::from_kind_and_msg(
+            SparseFormatErrorKind::InvalidStructure,
+            "Length of offset array is not equal to (major_dim + 1).",
+        ));
+    }
+
+    // Check that the first and last offsets conform to the specification
+    {
+        let first_offset_ok = *major_offsets.first().unwrap() == 0;
+        let last_offset_ok = *major_offsets.last().unwrap() == minor_indices.len();
+        if !first_offset_ok || !last_offset_ok {
+            return Err(SparseFormatError::from_kind_and_msg(
+                SparseFormatErrorKind::InvalidStructure,
+                "First or last offset is incompatible with format.",
+            ));
+        }
+    }
+
+    let mut minor_index_permutation: Vec<usize> = (0..minor_indices.len()).collect();
+
+    // Test that each lane has strictly monotonically increasing minor indices, i.e.
+    // minor indices within a lane are sorted, unique. Sort minor indices within a lane if needed.
+    // In addition, each minor index must be in bounds with respect to the minor dimension.
+    {
+        for lane_idx in 0..major_dim {
+            let range_start = major_offsets[lane_idx];
+            let range_end = major_offsets[lane_idx + 1];
+
+            // Test that major offsets are monotonically increasing
+            if range_start > range_end {
+                return Err(SparseFormatError::from_kind_and_msg(
+                    SparseFormatErrorKind::InvalidStructure,
+                    "Offsets are not monotonically increasing.",
+                ));
+            }
+
+            let minor_idx_in_lane = minor_indices.get(range_start..range_end).ok_or(
+                SparseFormatError::from_kind_and_msg(
+                    SparseFormatErrorKind::IndexOutOfBounds,
+                    "A major index is out of bounds.",
+                ),
+            )?;
+
+            // We test for in-bounds, uniqueness and monotonicity at the same time
+            // to ensure that we only visit each minor index once
+            let mut iter = minor_idx_in_lane.iter();
+            let mut prev = None;
+            let mut nonmonotonic = false;
+
+            while let Some(next) = iter.next().copied() {
+                if next >= minor_dim {
+                    return Err(SparseFormatError::from_kind_and_msg(
+                        SparseFormatErrorKind::IndexOutOfBounds,
+                        "A minor index is out of bounds.",
+                    ));
+                }
+
+                if let Some(prev) = prev {
+                    if prev == next {
+                        return Err(SparseFormatError::from_kind_and_msg(
+                            SparseFormatErrorKind::DuplicateEntry,
+                            "Input data contains duplicate entries.",
+                        ));
+                    }
+                    if prev > next {
+                        nonmonotonic = true;
+                    }
+                }
+                prev = Some(next);
+            }
+            if nonmonotonic && sort {
+                // sort permutation values for monotonic index order
+                minor_index_permutation[range_start..range_end].sort_by(|a, b| {
+                    let x = &minor_idx_in_lane[*a - range_start];
+                    let y = &minor_idx_in_lane[*b - range_start];
+                    x.partial_cmp(y).unwrap()
+                });
+            }
+        }
+    }
+
+    if sort {
+        let unsorted_indices: Vec<usize> = minor_indices.iter().map(|i| i.clone()).collect();
+        for (index, &offset) in minor_index_permutation.iter().enumerate() {
+            let u: usize = unsorted_indices[offset];
+            minor_indices[index] = u;
+        }
+        if let Some(values) = values {
+            if minor_indices.len() != values.len() {
+                return Err(SparseFormatError::from_kind_and_msg(
+                    SparseFormatErrorKind::InvalidStructure,
+                    "Number of values and minor indices must be the same",
+                ));
+            }
+            let unsorted_values: Vec<T> =
+                Vec::from_iter(((0..values.len()).map(|i| &values[i])).cloned());
+            for (index, &offset) in minor_index_permutation.iter().enumerate() {
+                let u: T = unsorted_values[offset].clone();
+                values[index] = u;
+            }
+        }
+    }
+    Ok(())
 }
